:sparkles: ci: triton inference server
